Extract `WrapBezier3d` and `WrapCurvePad` to `wrap_shape1d.cpp`
* Move the C++ wrapper classes `WrapBezier3d` and `WrapCurvePad` from file
  `wrap_World.cpp` to `wrap_shape1d.cpp`.

Claude code command:
* Extract `WrapBezier3d` and `WrapCurvePad` from the file `wrap_World.cpp` to
  `wrap_shape1d.cpp`
