+++ cpp/modmesh/universe/pymod/wrap_World.cpp
@@ -530,369 +530,6 @@ WrapTrianglePad<T> & WrapTrianglePad<T>::wrap_geometry()
 }
 
 template <typename T>
-class MODMESH_PYTHON_WRAPPER_VISIBILITY WrapBezier3d
-    : public WrapBase<WrapBezier3d<T>, Bezier3d<T>>
-{
-
-public:
-
-    using base_type = WrapBase<WrapBezier3d<T>, Bezier3d<T>>;
-    using wrapped_type = typename base_type::wrapped_type;
-
-    using point_type = typename wrapped_type::point_type;
-
-    friend typename base_type::root_base_type;
-
-protected:
-
-    WrapBezier3d(pybind11::module & mod, char const * pyname, char const * pydoc);
-
-    WrapBezier3d & wrap_management();
-    WrapBezier3d & wrap_accessor();
-    WrapBezier3d & wrap_geometry();
-}; /* end class WrapBezier3d */
-
-template <typename T>
-WrapBezier3d<T>::WrapBezier3d(pybind11::module & mod, const char * pyname, const char * pydoc)
-    : base_type(mod, pyname, pydoc)
-{
-    namespace py = pybind11;
-
-    (*this)
-        .wrap_management()
-        .wrap_accessor()
-        .wrap_geometry()
-        //
-        ;
-}
-
-template <typename T>
-WrapBezier3d<T> & WrapBezier3d<T>::wrap_management()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def(py::init<point_type const &, point_type const &, point_type const &, point_type const &>(),
-             py::arg("p0"),
-             py::arg("p1"),
-             py::arg("p2"),
-             py::arg("p3"))
-        .def(
-            "__repr__",
-            [](wrapped_type const & self)
-            {
-                static char const * btypename = std::is_same_v<T, double> ? "Bezier3dFp64" : "Bezier3dFp32";
-                static char const * ptypename = std::is_same_v<T, double> ? "Point3dFp64" : "Point3dFp32";
-                return std::format("{}({}({}), {}({}), {}({}), {}({}))",
-                                   btypename,
-                                   ptypename,
-                                   self.p0().value_string(),
-                                   ptypename,
-                                   self.p1().value_string(),
-                                   ptypename,
-                                   self.p2().value_string(),
-                                   ptypename,
-                                   self.p3().value_string());
-            })
-        .def_alias("__repr__", "__str__")
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapBezier3d<T> & WrapBezier3d<T>::wrap_accessor()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def("__len__",
-             [](wrapped_type const &)
-             { return 4; })
-        .def(
-            "__getitem__",
-            [](wrapped_type const & self, size_t it)
-            {
-                point_type ret;
-                switch (it)
-                {
-                case 0: ret = self.p0(); break;
-                case 1: ret = self.p1(); break;
-                case 2: ret = self.p2(); break;
-                case 3: ret = self.p3(); break;
-                default: throw std::out_of_range("Bezier3d: (control) i 4 >= size 4"); break;
-                }
-                return ret;
-            })
-        .def(
-            "__setitem__",
-            [](wrapped_type & self, size_t it, point_type const & p)
-            {
-                switch (it)
-                {
-                case 0: self.p0() = p; break;
-                case 1: self.p1() = p; break;
-                case 2: self.p2() = p; break;
-                case 3: self.p3() = p; break;
-                default: throw py::stop_iteration(); break;
-                }
-            })
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapBezier3d<T> & WrapBezier3d<T>::wrap_geometry()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def("sample", &wrapped_type::sample, py::arg("nlocus"))
-        .def(
-            "mirror",
-            [](wrapped_type & self, std::string const & axis)
-            {
-                if (axis == "x" || axis == "X")
-                {
-                    self.mirror(Axis::X);
-                }
-                else if (axis == "y" || axis == "Y")
-                {
-                    self.mirror(Axis::Y);
-                }
-                else if (axis == "z" || axis == "Z")
-                {
-                    self.mirror(Axis::Z);
-                }
-                else
-                {
-                    throw std::invalid_argument("Bezier3d::mirror: axis must be 'x', 'y', or 'z'");
-                }
-            },
-            py::arg("axis"))
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-class MODMESH_PYTHON_WRAPPER_VISIBILITY WrapCurvePad
-    : public WrapBase<WrapCurvePad<T>, CurvePad<T>, std::shared_ptr<CurvePad<T>>>
-{
-
-public:
-
-    using base_type = WrapBase<WrapCurvePad<T>, CurvePad<T>, std::shared_ptr<CurvePad<T>>>;
-    using wrapped_type = typename base_type::wrapped_type;
-
-    using value_type = typename base_type::wrapped_type::value_type;
-    using point_type = typename base_type::wrapped_type::point_type;
-    using segment_type = typename base_type::wrapped_type::segment_type;
-    using bezier_type = typename base_type::wrapped_type::bezier_type;
-    using point_pad_type = typename base_type::wrapped_type::point_pad_type;
-
-    friend typename base_type::root_base_type;
-
-protected:
-
-    WrapCurvePad(pybind11::module & mod, char const * pyname, char const * pydoc);
-
-    WrapCurvePad & wrap_management();
-    WrapCurvePad & wrap_accessor_bezier();
-    WrapCurvePad & wrap_accessor_point();
-    WrapCurvePad & wrap_geometry();
-}; /* end class WrapCurvePad */
-
-template <typename T>
-WrapCurvePad<T>::WrapCurvePad(pybind11::module & mod, const char * pyname, const char * pydoc)
-    : base_type(mod, pyname, pydoc)
-{
-    namespace py = pybind11;
-
-    (*this)
-        .wrap_management()
-        .wrap_accessor_bezier()
-        .wrap_accessor_point()
-        .wrap_geometry()
-        //
-        ;
-}
-
-template <typename T>
-WrapCurvePad<T> & WrapCurvePad<T>::wrap_management()
-{
-    namespace py = pybind11;
-
-    // Constructors
-    (*this)
-        .def(
-            py::init(
-                [](uint8_t ndim)
-                { return wrapped_type::construct(ndim); }),
-            py::arg("ndim"))
-        .def(
-            py::init(
-                [](uint8_t ndim, size_t nelem)
-                { return wrapped_type::construct(ndim, nelem); }),
-            py::arg("ndim"),
-            py::arg("nelem"))
-        //
-        ;
-
-    (*this)
-        .def_property_readonly("ndim", &wrapped_type::ndim)
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapCurvePad<T> & WrapCurvePad<T>::wrap_accessor_bezier()
-{
-    namespace py = pybind11;
-
-    // Python dunder accessor.
-    (*this)
-        .def("__len__", &wrapped_type::size)
-        .def("__getitem__", &wrapped_type::get_at)
-        .def("__setitem__", &wrapped_type::set_at)
-        //
-        ;
-
-    // Append Bezier3d
-    (*this)
-        .def(
-            "append",
-            [](wrapped_type & self, bezier_type const & c)
-            {
-                self.append(c);
-            },
-            py::arg("c"))
-        .def(
-            "append",
-            [](wrapped_type & self, point_type const & p0, point_type const & p1, point_type const & p2, point_type const & p3)
-            {
-                self.append(p0, p1, p2, p3);
-            },
-            py::arg("p0"),
-            py::arg("p1"),
-            py::arg("p2"),
-            py::arg("p3"))
-        //
-        ;
-
-    // Additional getter and setter for Bezier3d.
-    (*this)
-        .def("get_at", &wrapped_type::get_at)
-        .def("set_at", &wrapped_type::set_at)
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapCurvePad<T> & WrapCurvePad<T>::wrap_accessor_point()
-{
-    namespace py = pybind11;
-
-    // x, y, z, point element accessors.
-#define DECL_WRAP(NAME) \
-    .def(#NAME, [](wrapped_type const & self, size_t i) { return self.NAME(i); })
-    // clang-format off
-    (*this)
-        DECL_WRAP(x0_at)
-        DECL_WRAP(y0_at)
-        DECL_WRAP(z0_at)
-        DECL_WRAP(x1_at)
-        DECL_WRAP(y1_at)
-        DECL_WRAP(z1_at)
-        DECL_WRAP(x2_at)
-        DECL_WRAP(y2_at)
-        DECL_WRAP(z2_at)
-        DECL_WRAP(x3_at)
-        DECL_WRAP(y3_at)
-        DECL_WRAP(z3_at)
-        DECL_WRAP(p0_at)
-        DECL_WRAP(p1_at)
-        DECL_WRAP(p2_at)
-        DECL_WRAP(p3_at)
-        ;
-    // clang-format on
-#undef DECL_WRAP
-
-    // x, y, z, point array/batch accessors.
-#define DECL_WRAP(NAME)         \
-    .def_property_readonly(     \
-        #NAME,                  \
-        [](wrapped_type & self) \
-        { return self.NAME(); })
-    // clang-format off
-    (*this)
-        DECL_WRAP(x0)
-        DECL_WRAP(y0)
-        DECL_WRAP(z0)
-        DECL_WRAP(x1)
-        DECL_WRAP(y1)
-        DECL_WRAP(z1)
-        DECL_WRAP(x2)
-        DECL_WRAP(y2)
-        DECL_WRAP(z2)
-        DECL_WRAP(x3)
-        DECL_WRAP(y3)
-        DECL_WRAP(z3)
-        DECL_WRAP(p0)
-        DECL_WRAP(p1)
-        DECL_WRAP(p2)
-        DECL_WRAP(p3)
-        ;
-    // clang-format on
-#undef DECL_WRAP
-
-    return *this;
-}
-
-template <typename T>
-WrapCurvePad<T> & WrapCurvePad<T>::wrap_geometry()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def("sample", &wrapped_type::sample, py::arg("length"))
-        .def(
-            "mirror",
-            [](wrapped_type & self, std::string const & axis)
-            {
-                if (axis == "x" || axis == "X")
-                {
-                    self.mirror(Axis::X);
-                }
-                else if (axis == "y" || axis == "Y")
-                {
-                    self.mirror(Axis::Y);
-                }
-                else if (axis == "z" || axis == "Z")
-                {
-                    self.mirror(Axis::Z);
-                }
-                else
-                {
-                    throw std::invalid_argument("CurvePad::mirror: axis must be 'x', 'y', or 'z'");
-                }
-            },
-            py::arg("axis"))
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
 class MODMESH_PYTHON_WRAPPER_VISIBILITY WrapBoundBox3d
     : public WrapBase<WrapBoundBox3d<T>, BoundBox3d<T>>
 {
@@ -1166,10 +803,6 @@ void wrap_World(pybind11::module & mod)
     WrapTriangle3d<double>::commit(mod, "Triangle3dFp64", "Triangle3dFp64");
     WrapTrianglePad<float>::commit(mod, "TrianglePadFp32", "TrianglePadFp32");
     WrapTrianglePad<double>::commit(mod, "TrianglePadFp64", "TrianglePadFp64");
-    WrapBezier3d<float>::commit(mod, "Bezier3dFp32", "Bezier3dFp32");
-    WrapBezier3d<double>::commit(mod, "Bezier3dFp64", "Bezier3dFp64");
-    WrapCurvePad<float>::commit(mod, "CurvePadFp32", "CurvePadFp32");
-    WrapCurvePad<double>::commit(mod, "CurvePadFp64", "CurvePadFp64");
     WrapWorld<float>::commit(mod, "WorldFp32", "WorldFp32");
     WrapWorld<double>::commit(mod, "WorldFp64", "WorldFp64");
 }
